fix: search by tags with trailing hypens or spaces
This commit normalizes tags in the url to never have trailing hyphens. Then, a regular expression is used to find all articles that match the tag. Matches allow for any number and combination of spaces and hyphens between the tag's characters. For example, "President davis", "president-davis", "President-Davis-  ", and "presidentdavis" all resolve to the same list of articles